Quality of life updates.
Changes on 26: should cancel the picking if user clicks any other mouse button other than left.

Changes on 63: added check to prevent self-referencing picking

Changes on 74: Main one! Replaced hardcoding of GameObject type to the actual type of the property field. So now by just adding more "[CustomPropertyDrawer(typeof( ))]"to the class it is possible to add those pickers to any field type that deals with picking game objects or MonoBehaviors on the objects.

Changes on 76: Added condition that will change the button's color to indicate it's in "pick object" mode.
Changes on 80: Just added a helpful tooltip.